# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenOutlier API
  version: 1.0.0
  description: API for tracking channels, scanning niches, finding outliers, and saving references.
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
security:
  - ApiKeyAuth: []
paths:
  /api/health:
    get:
      security: []
      summary: Health check
      responses:
        "200":
          description: OK
  /api/collections:
    get:
      summary: List collections
      responses:
        "200":
          description: Collections returned
    post:
      summary: Create collection
      responses:
        "201":
          description: Collection created
  /api/collections/{id}:
    get:
      summary: Get collection
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Collection returned
  /api/tracked-channels:
    get:
      summary: List tracked channels
      responses:
        "200":
          description: Tracked channels returned
    post:
      summary: Add tracked channel
      responses:
        "201":
          description: Tracked channel added
  /api/tracked-channels/discover:
    post:
      summary: Discover channels to track
      responses:
        "200":
          description: Suggestions returned
  /api/scan:
    post:
      summary: Trigger scan
      responses:
        "202":
          description: Scan started
  /api/scan/status:
    get:
      summary: Get scan status
      responses:
        "200":
          description: Scan status returned
  /api/discover/outliers:
    get:
      summary: Search outlier feed
      responses:
        "200":
          description: Outliers returned
  /api/discover/similar-topics:
    get:
      summary: Get similar topics for a video
      responses:
        "200":
          description: Similar topics returned
  /api/discover/niches:
    get:
      summary: Get niche clusters
      responses:
        "200":
          description: Niche data returned
  /api/collections/{id}/references/search:
    post:
      summary: Search references within a collection
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Search results returned
  /api/collections/{id}/references:
    get:
      summary: List saved references
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: References returned
    post:
      summary: Save reference
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Reference saved
  /api/collections/{id}/references/{referenceId}:
    delete:
      summary: Remove one saved reference
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: referenceId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Reference removed
  /api/collections/{id}/export:
    get:
      summary: Export a collection
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: format
          required: false
          schema:
            type: string
            enum: [json, csv]
      responses:
        "200":
          description: Collection export returned
  /api/collections/{id}/references/import-video:
    post:
      summary: Import one video as a reference
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Video imported